+++ mod-source-record-storage-server/src/main/resources/templates/db_scripts/migration_scripts/records_delete_matched_profile_id.sql
@@ -1,0 +1,3 @@
+-- delete "matchedProfileId" field
+UPDATE ${myuniversity}_${mymodule}.records
+SET jsonb = jsonb - 'matchedProfileId';